Update gMSA statemanager tests
